requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
pandas==2.2.3
google-genai>=1.0.0
pydantic==2.10.4
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
